fix: moved endpoint-related classes to corresponding modules
