+++ aidial_sdk/utils/pydantic.py
@@ -1,0 +1,6 @@
+from aidial_sdk.pydantic_v1 import BaseModel
+
+
+class ExtraForbidModel(BaseModel):
+    class Config:
+        extra = "forbid"
