+++ tests/utils/errors.py
@@ -1,4 +1,4 @@
-from pydantic import BaseModel
+from aidial_sdk.pydantic_v1 import BaseModel
 
 
 class Error(BaseModel):
